% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{applications}
\alias{applications}
\title{Function to create data frames for applications received. Supports 'All' departments.}
\usage{
applications(data, dates_obj)
}
\arguments{
\item{data}{The list of data frames resulting from the HR_pulldata() function}

\item{dates_obj}{The list of dates resulting from HR_dates() function}
}
\value{
A list of data frames
}
\description{
Function to create data frames for applications received. Supports 'All' departments.
}
\examples{
dept_applications <- applications(dept_HRdata, dates)
}
