% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{applications_plot}
\alias{applications_plot}
\title{Function to generate plot for applications time series}
\usage{
applications_plot(data, pal)
}
\arguments{
\item{data}{The list of data frames resulting from the applications() function}

\item{pal}{Palette of colors, usually resulting from kpiPal() function}
}
\value{
A ggplot object
}
\description{
Function to generate plot for applications time series
}
\examples{
print(applications_plot(dept_applications, pal))
}
