% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{applications_table}
\alias{applications_table}
\title{Function to generate a table of applications received in the reporting month or reporting year}
\usage{
applications_table(data, rpt_date, period = "month")
}
\arguments{
\item{data}{The list of data frames resulting from the applications() function}

\item{rpt_date}{A date during the reporting month}

\item{period}{To show either reporting month or reporting year. Must be either 'month' or 'year'}
}
\value{
A kable object
}
\description{
Function to generate a table of applications received in the reporting month or reporting year
}
\examples{
print(applications_table(dept_applications, dates$reporting_smonth, period = 'month'))
print(applications_table(dept_applications, dates$reporting_smonth, period = 'year'))
}
