% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{applications_text}
\alias{applications_text}
\title{Function to generate reporting month and year to date text for application section of report}
\usage{
applications_text(data, rpt_date, seasonal = FALSE)
}
\arguments{
\item{data}{The list of data frames resulting from the applications() function}

\item{rpt_date}{Date in reporting month}

\item{seasonal}{TRUE or FALSE binary to compare hiring numbers to previous month (FALSE) or same month of the previous year (TRUE). Defaults to FALSE.}
}
\value{
A list of character values
}
\description{
Function to generate reporting month and year to date text for application section of report
}
\examples{
dept_apps_text <- applications_text(dept_applications, dates$reporting_smonth)
}
