% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{current_employees}
\alias{current_employees}
\title{Function to pull categorize employees by status}
\usage{
current_employees(data, include_seasonals = FALSE)
}
\arguments{
\item{data}{The data frames resulting from the HR_pulldata() function}

\item{include_seasonals}{TRUE or FALSE binary to include seasonal and temporary employees. Defaults to FALSE.}
}
\value{
A dataframe with employee counts by category that can be used in the current_employees_text() and current_employees_table() functions
}
\description{
Function to pull categorize employees by status
}
\examples{
dept_employees_cat <- current_employees(dept_HRdata, include_seasonals = FALSE)
}
