% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{current_employees_table}
\alias{current_employees_table}
\title{Function to make table of current employee categories}
\usage{
current_employees_table(data, include_seasonals = FALSE)
}
\arguments{
\item{data}{The employee_categories data frame resulting from current_employees() function}

\item{include_seasonals}{TRUE or FALSE binary to include seasonal and temporary employees. Defaults to FALSE.}
}
\value{
A kable table with employee counts by category
}
\description{
Function to make table of current employee categories
}
\examples{
print(current_employees_table(dept_employees_cat$employee_categories, include_seasonals = FALSE))
}
