% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{current_employees_text}
\alias{current_employees_text}
\title{Function to create header and explanatory text for current employees section of HR report}
\usage{
current_employees_text(data, dept_name, rpt_enddate, include_seasonals = FALSE)
}
\arguments{
\item{data}{The data frames resulting from current_employees() function}

\item{dept_name}{Name of department(s) as it should appear in text}

\item{rpt_enddate}{Last day in reporting month (produced by HR_dates() function)}

\item{include_seasonals}{TRUE or FALSE binary to include seasonal and temporary employees. Defaults to FALSE.}
}
\value{
A character value
}
\description{
Function to create header and explanatory text for current employees section of HR report
}
\examples{
dept_employees_txt <- current_employees_text(dept_employees_cat, "Health Department", dates$reporting_emonth, include_seasonals = FALSE)
}
