% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{gen_date_readable}
\alias{gen_date_readable}
\title{Function to generate the proper date generated text.}
\usage{
gen_date_readable(page = "Dept")
}
\arguments{
\item{page}{Page type of rendered text, default = "Dept" which includes entire text, any other value will just provide the Month, Day and Year as text.}
}
\value{
HTML Date Rendered Text
}
\description{
Function to generate the proper date generated text.
}
\examples{
\dontrun{

```{r results='asis'}
gen_date_readable()
```
}
}
