% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{hires}
\alias{hires}
\title{Function to create data frames for hiring}
\usage{
hires(data, jde_dept, dates_obj, seasonal_rehires = FALSE)
}
\arguments{
\item{data}{The list of data frames resulting from the HR_pulldata() function}

\item{jde_dept}{Name of the department in JDE (recommended to set as variable). Supports 'All' departments.}

\item{dates_obj}{The list of dates resulting from HR_dates() function}

\item{seasonal_rehires}{TRUE or FALSE binary to produce additional data frame with seasonal rehires time series. Defaults to FALSE.}
}
\value{
A list of data frames
}
\description{
Function to create data frames for hiring
}
\examples{
dept_hires <- hires(dept_HRdata, jde_dept, dates, seasonal_rehires = FALSE)
}
