% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{hires_plot}
\alias{hires_plot}
\title{Function to generate plot(s) of hiring time series}
\usage{
hires_plot(data, pal, seasonal_rehires = FALSE)
}
\arguments{
\item{data}{The list of data frames resulting from the hires() function}

\item{pal}{Palette of colors, usually resulting from kpiPal() function}

\item{seasonal_rehires}{TRUE or FALSE binary to produce additional ggplot of seasonal rehires time series. Defaults to FALSE.}
}
\value{
A single ggplot or list of ggplot objects
}
\description{
Function to generate plot(s) of hiring time series
}
\examples{
print(hires_plot(dept_hires, pal, seasonal_rehires = FALSE))
}
