% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{hires_text}
\alias{hires_text}
\title{Function to generate reporting month and year to date text for hiring section of report}
\usage{
hires_text(data, rpt_date, seasonal = FALSE, seasonal_rehires = FALSE)
}
\arguments{
\item{data}{The list of data frames resulting from the hires() function}

\item{rpt_date}{Date in reporting month}

\item{seasonal}{TRUE or FALSE binary to compare hiring numbers to previous month (FALSE) or same month of the previous year (TRUE). Defaults to FALSE.}

\item{seasonal_rehires}{TRUE or FALSE binary to add rehires numbers to header. Defaults to FALSE.}
}
\value{
A list of character values
}
\description{
Function to generate reporting month and year to date text for hiring section of report
}
\examples{
dept_hire_text <- hires_text(dept_hires, dates$reporting_smonth, seasonal = FALSE, seasonal_rehires = FALSE)
}
