% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{HR_dates}
\alias{HR_dates}
\title{Function to create all dates necessary for all HR functions. Use this as the dates object in all functions that call for it}
\usage{
HR_dates()
}
\value{
A list of four dates -- reporting month floor date, reporting month ceiling date, reporting period floor date (start of previous year) and reporting period end date
}
\description{
Function to create all dates necessary for all HR functions. Use this as the dates object in all functions that call for it
}
\examples{
dates <- HR_dates()
}
