% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{HR_pulldata}
\alias{HR_pulldata}
\title{Function to pull all HR-related data. Before running this function, you must have a warehouse connection and the name of the department as it is listed in JDE, OnBase, and Novatime (if pulling for a specific department)}
\usage{
HR_pulldata(
  wh_con,
  jde_dept = "All",
  onbase_dept = "All",
  novatime_dept = "All",
  rpt_enddate,
  include_seasonals = FALSE
)
}
\arguments{
\item{wh_con}{Data warehouse connection}

\item{jde_dept}{Name of department as it is listed in JDE. Defaults to 'All'.}

\item{onbase_dept}{Name of department as it is listed in OnBase. Defaults to 'All'.}

\item{novatime_dept}{Name of department as it is listed in Novatime. Defaults to 'All'.}

\item{rpt_enddate}{reporting_emonth in HR_dates() function}

\item{include_seasonals}{TRUE or FALSE binary to include seasonal and temporary employees. Defaults to FALSE.}
}
\value{
A list of dataframes containing all HR-related data
}
\description{
Function to pull all HR-related data. Before running this function, you must have a warehouse connection and the name of the department as it is listed in JDE, OnBase, and Novatime (if pulling for a specific department)
}
\examples{
dept_HRdata <- HR_pulldata(wh_con, jde_dept, onbase_dept, novatime_dept, rpt_enddate = dates$reporting_emonth, include_seasonals = FALSE)
all_HRdata <- HR_pulldata(wh_con, rpt_enddate = dates$reporting_emonth, include_seasonals = FALSE)
}
