% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{kpi_callout}
\alias{kpi_callout}
\title{Function to build the KPI callout bar for Quarto Sites. This function should be inside an 'asis' code chunk.}
\usage{
kpi_callout(type, kpi, text)
}
\arguments{
\item{type}{What type of callout found at: https://quarto.org/docs/authoring/callouts.html}

\item{kpi}{Callout/KPI Title}

\item{text}{Content of the callout box}
}
\value{
Quarto Callout Markdown
}
\description{
Function to build the KPI callout bar for Quarto Sites. This function should be inside an 'asis' code chunk.
}
\examples{
\dontrun{

```{r results='asis'}
total_pop <- nrow(filter(penguins, year == max(penguins$year)))
type <- if_else(total_pop < 200, "important", "tip")
kpi_callout(type, "Penguin Population", paste("The Penguin Population for 2009 was", total_pop))
```
}
}
