% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_general.R
\name{kpi_compare}
\alias{kpi_compare}
\title{A function to obtain current month, previous month, and same month of previous year values, or current year-to-date and the same period of the previous year values and calculate percent changes.}
\usage{
kpi_compare(data, date_col, rpt_date, method = "tally", val_col, mode = "my")
}
\arguments{
\item{data}{A data frame.}

\item{date_col}{The date column in the data dataframe.}

\item{rpt_date}{The first day of the month of the report (this will serve as the base month in the comparisons).}

\item{method}{The method by which you want to process the data. It has four values "tally" counts rows, "sum" sums values in a specified column, "average" averages values in a specified column, and "distinct" counts the distinct values in a specified column.}

\item{val_col}{Only necessary for methods "sum", "average", and "distinct". The column which has the values you want to perform that function on.}

\item{mode}{Can be either "my" or "ytd". The former calculates month-over-month and year-over-year change, the latter calculates current change between current year-to-date and the previous year over the same period.}
}
\value{
A list with the values being compared and the comparisons.
}
\description{
A function to obtain current month, previous month, and same month of previous year values, or current year-to-date and the same period of the previous year values and calculate percent changes.
}
\examples{
metric_cmp <- kpi_compare(metric_data, date, rpt_date, mode = "ytd")
metric_cmp <- kpi_compare(metric_data, date, rpt_date, method = "avg", val_col = totals)
}
