% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_general.R
\name{kpi_header_text}
\alias{kpi_header_text}
\title{A function that creates the text at the top of each KPI showing the value and change from the previous specified period}
\usage{
kpi_header_text(kpi_cmp, rpt_date, prior_val = "month", metric_txt)
}
\arguments{
\item{kpi_cmp}{A kpi_compare list produced by kpi_compare or kpi_pcts.}

\item{rpt_date}{The first day of the month of the report.}

\item{prior_val}{The value you want to compare to "month" for previous month, "year" for the same month of the previous year, or "ytd" to compare the current year-to-date with the same period of the previous year.}

\item{metric_txt}{The text you want to appear after the reporting month count.}
}
\value{
A string with R markdown formatting.
}
\description{
A function that creates the text at the top of each KPI showing the value and change from the previous specified period
}
\examples{
metric_header_txt <- kpi_header_text(metric_cmp, rpt_date, prior_val = "year", metric_txt = "calls to IT help desk")
}
