% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_general.R
\name{kpi_pcts}
\alias{kpi_pcts}
\title{Function to calculate month-over-month and year-over-year or year-to-date percentage change.}
\usage{
kpi_pcts(rm_cnt, pm_cnt, py_cnt, mode = "my")
}
\arguments{
\item{rm_cnt}{Reporting month count in "my" mode and current year-to-date count in "ytd" mode.}

\item{pm_cnt}{Previous month count in "my" mode or previous year-to-date count in "ytd" mode.}

\item{py_cnt}{Same month of previous year count (this is optional and only needed in my mode)}

\item{mode}{Can be either "my" or "ytd". The former calculates month-over-month and year-over-year change, the latter calculates change between current year-to-date and the previous year over the same period.}
}
\value{
A list with the entered values, the percent change, and up or down arrow, whether it increased or decreased, and the absolute change
}
\description{
Function to calculate month-over-month and year-over-year or year-to-date percentage change.
}
\examples{
metric_pcts <- kpi_pcts(metric_cnt, metric_pm_cnt, metric_py_cnt)
}
