% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{kpi_trend_plot}
\alias{kpi_trend_plot}
\title{A function that creates a trend plot, showing monthly data for the current year and four previous years, with the current year highlighted. It must be followed with labs and theme arguments.}
\usage{
kpi_trend_plot(
  data,
  cnt_col,
  month_lag = 0,
  rpt_month = rpt_date,
  date_col = rpting_date,
  palette = pal
)
}
\arguments{
\item{data}{A data frame.}

\item{cnt_col}{The column with the counts or values you want to plot.}

\item{month_lag}{How many months you want it to lag from the report date (default 0).}

\item{rpt_month}{The first date of the month of the KPI report (default rpt_date).}

\item{date_col}{The date column in the data frame (default rpting_date).}

\item{palette}{The palette you want it to use (default pal).}
}
\value{
A ggplot object.
}
\description{
A function that creates a trend plot, showing monthly data for the current year and four previous years, with the current year highlighted. It must be followed with labs and theme arguments.
}
\examples{
metric_data \%>\% kpi_trend_plot(totals) + labs(title = 'Plot title', ylab = 'Count'))
metric_data \%>\% kpi_trend_plot(totals, month_lag = 3, date_col = date) + labs(title = 'Plot title', ylab = 'Count') + theme_minimal())
}
