% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{kpi_trend_plotly}
\alias{kpi_trend_plotly}
\title{A function that creates a trend plot (ggplot), showing monthly data for the current year and four previous years, with the current year highlighted. It must be followed with labs and theme arguments. This function additionally includes a text argument, which will create the tooltip in ggplotly.}
\usage{
kpi_trend_plotly(
  data,
  cnt_col,
  month_lag = 0,
  rpt_month = rpt_date,
  date_col = rpting_date,
  palette = pal,
  text_custom = FALSE,
  metric_text
)
}
\arguments{
\item{data}{A data frame.}

\item{cnt_col}{The column with the counts or values you want to plot.}

\item{month_lag}{How many months you want it to lag from the report date (default 0).}

\item{rpt_month}{The first date of the month of the KPI report (default rpt_date).}

\item{date_col}{The date column in the data frame (default rpting_date).}

\item{palette}{The palette you want it to use (default pal).}

\item{text_custom}{A true/false binary for showing the standard metric text (default) or creating a custom tooltop}

\item{metric_text}{The descriptive text for the cnt_col or custom paste function (depending on text_custom) that will appear in the tooltip}
}
\value{
A ggplot object.
}
\description{
A function that creates a trend plot (ggplot), showing monthly data for the current year and four previous years, with the current year highlighted. It must be followed with labs and theme arguments. This function additionally includes a text argument, which will create the tooltip in ggplotly.
}
\examples{
census_plot <- kpi_trend_plotly(monthly_avg_census, avg_census, rpt_month = reporting_mo, date_col = census_month, palette = pal, metric_text = "average daily residents") + labs(title = "Monthly Average Daily Census", y = "Residents") + kpi_trend_theme
}
