% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{kpi_trend_theme}
\alias{kpi_trend_theme}
\title{A theme for KPI trend plots}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 94.
}
\usage{
kpi_trend_theme
}
\value{
A ggplot object.
}
\description{
A theme for KPI trend plots
}
\examples{
plot + labs(title = "Plot title", y = "Count") + kpi_trend_theme())
}
\keyword{datasets}
