% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{kpiPal}
\alias{kpiPal}
\title{A function that creates a uniform color scheme for KPI reports.}
\usage{
kpiPal(end_date)
}
\value{
A palette.
}
\description{
A function that creates a uniform color scheme for KPI reports.
}
\examples{
palette = kpiPal(rpt_date)
}
