% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{net}
\alias{net}
\title{Function to create data frames for net employees. Does not support 'All' departments.}
\usage{
net(
  hires_obj,
  terminations_df,
  transfers_df,
  dates_obj,
  seasonal_rehires = FALSE
)
}
\arguments{
\item{hires_obj}{The list of data frames resulting from the hires() function}

\item{terminations_df}{The terminations_time data frame produced by the terminations() function}

\item{transfers_df}{The transfers_time data frame produced by the transfers() function}

\item{dates_obj}{The list of dates resulting from HR_dates() function}
}
\value{
A list of data frames
}
\description{
Function to create data frames for net employees. Does not support 'All' departments.
}
\examples{
dept_net <- net(dept_hires, dept_terminations$terminations_time, dept_transfers$transfers_time, dates)
}
