% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{net_plot}
\alias{net_plot}
\title{Function to generate plot for title change time series}
\usage{
net_plot(data, seasonal_rehires = FALSE)
}
\arguments{
\item{data}{The list of data frames resulting from the net() function}

\item{seasonal_rehires}{A TRUE or FALSE binary that changes the caption of the ggplot. Defaults to FALSE.}
}
\value{
A ggplot object
}
\description{
Function to generate plot for title change time series
}
\examples{
print(net_plot(dept_net, seasonal_rehires = FALSE))
}
