% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{net_text}
\alias{net_text}
\title{Function to generate reporting year to date text for the net employees section of report}
\usage{
net_text(data, rpt_date)
}
\arguments{
\item{data}{The list of data frames resulting from the net() function}

\item{rpt_date}{Date in reporting month}
}
\value{
A character value
}
\description{
Function to generate reporting year to date text for the net employees section of report
}
\examples{
dept_net_text <- net_text(dept_net, dates$reporting_smonth)
}
