% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotly_stack_fix}
\alias{plotly_stack_fix}
\title{A function to fix stacked bar charts so they reposition themselves when you deselect legend items.}
\usage{
plotly_stack_fix(plot)
}
\arguments{
\item{plot}{A ggplot object, likely created using kpi_trend_plotly}
}
\value{
A ggplotly object
}
\description{
A function to fix stacked bar charts so they reposition themselves when you deselect legend items.
}
\examples{
plotly_stack_fix(census_plot)
}
