% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotly_theme}
\alias{plotly_theme}
\title{A function to turn a kpi_trend_plotly object into a ggplotly object. Note: You cannot have subtitles in ggplotly.}
\usage{
plotly_theme(plot)
}
\arguments{
\item{plot}{A ggplot object, likely created using kpi_trend_plotly}
}
\value{
A ggplotly object
}
\description{
A function to turn a kpi_trend_plotly object into a ggplotly object. Note: You cannot have subtitles in ggplotly.
}
\examples{
plotly_theme(census_plot)
}
