% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTK.R
\name{rtk_chart}
\alias{rtk_chart}
\title{Function that filters RTK data, cleaned and formatted by rtk_data function, down to a specified date range, then creates the selected kable or ggplot object.}
\usage{
rtk_chart(data, rpt_sdate, palette, type = "type", range = "month")
}
\arguments{
\item{data}{Data frame of clean and formatted RTK data}

\item{rpt_sdate}{A date value that represents the first day of the reporting month}

\item{palette}{Color palette for time series}

\item{type}{Type of kable or gglot object returned. Must be 'type', 'close', 'late', 'table', 'status', or 'received'. 'Status' and 'received' types facet by department if more than one is present in data. Data format must match type selected.}

\item{range}{Date range displayed in plot or table. Must be 'month', '12months', 'timeseries', or 'ytd'.}
}
\value{
A kable or ggplot object
}
\description{
Function that filters RTK data, cleaned and formatted by rtk_data function, down to a specified date range, then creates the selected kable or ggplot object.
}
\examples{
\dontrun{
as_plots <- rtk_chart(as_transform$late, start_last_month, pal, type = "late", range = "timeseries")
print(as_plots)
ems_plots <- rtk_chart(ems_transform$status, start_last_month, kpipalette, type = "status", range = "12months")
print(ems_plots)
}
}
