% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTK.R
\name{rtk_data}
\alias{rtk_data}
\title{Function that manipulates RTK data, retrieved using function rtk_pull, into a list of data frames that can be used to create RTK KPI-related text, tables, charts, and graphs. Can be used with either 'single' or 'multi' data.}
\usage{
rtk_data(data, rpt_sdate, rpt_edate)
}
\arguments{
\item{data}{data frame of RTK data}

\item{rpt_sdate}{a date value that represents the first day of the reporting month}

\item{rpt_edate}{a date value that represents the last day of the reporting month}
}
\value{
A list of data frames of aggregated RTK data for the department(s) selected
}
\description{
Function that manipulates RTK data, retrieved using function rtk_pull, into a list of data frames that can be used to create RTK KPI-related text, tables, charts, and graphs. Can be used with either 'single' or 'multi' data.
}
\examples{
\dontrun{
as_transform <- rtk_data(data_as, start_last_month, end_last_month)
}
}
