% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTK.R
\name{rtk_pull}
\alias{rtk_pull}
\title{Function to pull RTK data for a particular department. If a department includes multiple divisions or if the report requires multiple departments to be shown, this distinction can be made.}
\usage{
rtk_pull(con, dept_list, col_list, mode = "single")
}
\arguments{
\item{con}{Data warehouse connection}

\item{dept_list}{A string containing the department name as it appears in GovQA. If there are multiple departments/divisions, each value must be surrounded by singular quotation marks and separated by a comma.}

\item{col_list}{A string containing the department-specific columns that show information type in GovQA. If there are multiple, they must be separated by a comma.}

\item{mode}{Can either be "single" or "multi" and defaults to "single". Use "multi" if there are multiple departments/divisions.}
}
\value{
A data frame of RTK requests for the specified department
}
\description{
Function to pull RTK data for a particular department. If a department includes multiple divisions or if the report requires multiple departments to be shown, this distinction can be made.
}
\examples{
\dontrun{
dept_list <- "'Administrative Services', 'Elections', 'Property Assessment', 'Purchasing', 'Real Estate'"
col_list <- "department_information_type_administrative_servi, department_information_type_elections, department_information_type_property_assessment, department_information_type_purchasing, department_information_type_real_estate"
data_as <- rtk_pull(wh_con, dept_list, col_list, mode = "multi")
data_ems <- rtk_pull(wh_con, "'Emergency Services'", "department_information_type_emergency_services", mode = "single")
}
}
