% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTK.R
\name{rtk_summary}
\alias{rtk_summary}
\title{Function that manipulates RTK data, retrieved using function rtk_pull, into a list of values that can be used to create RTK KPI-related text. Can be used with either 'single' or 'multi' data, but 'multi' data will be aggregated.}
\usage{
rtk_summary(data, rpt_sdate, rpt_edate)
}
\arguments{
\item{data}{data frame of RTK data}

\item{rpt_sdate}{a date value that represents the first day of the reporting month}

\item{rpt_edate}{a date value that represents the last day of the reporting month}
}
\value{
A named list of values of RTK KPIs
}
\description{
Function that manipulates RTK data, retrieved using function rtk_pull, into a list of values that can be used to create RTK KPI-related text. Can be used with either 'single' or 'multi' data, but 'multi' data will be aggregated.
}
\examples{
\dontrun{
as_summary <- rtk_data(data_as, start_last_month, end_last_month)
}
}
