% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTK.R
\name{rtk_text}
\alias{rtk_text}
\title{Function that uses RTK data, retrieved using function rtk_pull, to run rtk_summary which retrieves a list of values that are then used to create the text in the departmental RTK report. Can be used with either 'single' or 'multi' data, but 'multi' data will be aggregated.}
\usage{
rtk_text(data, rpt_sdate, rpt_edate, dept_text)
}
\arguments{
\item{data}{data frame of RTK data}

\item{rpt_sdate}{a date value that represents the first day of the reporting month}

\item{rpt_edate}{a date value that represents the last day of the reporting month}

\item{dept_text}{a text value that represents how the department or departments will be referred to within the text blocks of the report}
}
\value{
A named list of text strings for the departmental KPI report
}
\description{
Function that uses RTK data, retrieved using function rtk_pull, to run rtk_summary which retrieves a list of values that are then used to create the text in the departmental RTK report. Can be used with either 'single' or 'multi' data, but 'multi' data will be aggregated.
}
\examples{
\dontrun{
as_text <- rtk_text(data_as, start_last_month, end_last_month, "Administrative Services and its related divisions")
ems_text <- rtk_text(data_ems, start_last_month, end_last_month, "Emergency Services")
}
}
