% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{terminations_plot}
\alias{terminations_plot}
\title{Function to generate plot for termination time series}
\usage{
terminations_plot(data, pal)
}
\arguments{
\item{data}{The list of data frames resulting from the hires() function}

\item{pal}{Palette of colors, usually resulting from kpiPal() function}
}
\value{
A ggplot object
}
\description{
Function to generate plot for termination time series
}
\examples{
print(terminations_plot(dept_terminations, pal))
}
