% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{terminations_table}
\alias{terminations_table}
\title{Function to generate tables of termination reasons for reporting month and year to date}
\usage{
terminations_table(data, rpt_date)
}
\arguments{
\item{data}{The list of data frames resulting from the hires() function}

\item{rpt_date}{Date in reporting month}
}
\value{
A list of kable objects
}
\description{
Function to generate tables of termination reasons for reporting month and year to date
}
\examples{
print(terminations_table(dept_terminations, dates$reporting_smonth)$month_termreason_table)
print(terminations_table(dept_terminations, dates$reporting_smonth)$ytd_termreason_table)
}
