% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{time_off}
\alias{time_off}
\title{Function to create data frames for sick and vacation time taken. Supports 'All' departments.}
\usage{
time_off(data, dates_obj)
}
\arguments{
\item{data}{The list of data frames resulting from the HR_pulldata() function}

\item{dates_obj}{The list of dates resulting from HR_dates() function}
}
\value{
A list of data frames
}
\description{
Function to create data frames for sick and vacation time taken. Supports 'All' departments.
}
\examples{
dept_timeoff <- time_off(dept_HRdata, dates)
}
