% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{time_off_plots}
\alias{time_off_plots}
\title{Function to generate plots for sick and vacation time time series}
\usage{
time_off_plots(data, pal)
}
\arguments{
\item{data}{The list of data frames resulting from the time_off() function}

\item{pal}{Palette of colors, usually resulting from kpiPal() function}
}
\value{
A list of ggplot objects
}
\description{
Function to generate plots for sick and vacation time time series
}
\examples{
print(time_off_plots(dept_timeoff, pal)$sick_plot)
print(time_off_plots(dept_timeoff, pal)$vac_plot)
}
