% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{time_off_text}
\alias{time_off_text}
\title{Function to generate reporting month text for time off section of report}
\usage{
time_off_text(data, rpt_date)
}
\arguments{
\item{data}{The list of data frames resulting from the time_off() function}

\item{rpt_date}{Date in reporting month}
}
\value{
A list of character values
}
\description{
Function to generate reporting month text for time off section of report
}
\examples{
dept_timeoff_text <- time_off_text(dept_timeoff, dates$reporting_smonth)
}
