% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{title_change_text}
\alias{title_change_text}
\title{Function to generate reporting year to date text for title change section of report}
\usage{
title_change_text(data, rpt_date)
}
\arguments{
\item{data}{The list of data frames resulting from the title_changes() function}

\item{rpt_date}{Date in reporting month}
}
\value{
A list of character values
}
\description{
Function to generate reporting year to date text for title change section of report
}
\examples{
dept_titlechange_text <- title_change_text(dept_titlechanges, dates$reporting_smonth)
}
