% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{title_changes_plot}
\alias{title_changes_plot}
\title{Function to generate plot for title change time series}
\usage{
title_changes_plot(data, pal)
}
\arguments{
\item{data}{The list of data frames resulting from the title_changes() function}

\item{pal}{Palette of colors, usually resulting from kpiPal() function}
}
\value{
A ggplot object
}
\description{
Function to generate plot for title change time series
}
\examples{
print(title_changes_plot(dept_titlechanges, pal))
}
