% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{transfers}
\alias{transfers}
\title{Function to create data frames for transfers}
\usage{
transfers(data, jde_dept, dates_obj)
}
\arguments{
\item{data}{The list of data frames resulting from the HR_pulldata() function}

\item{jde_dept}{Name of the department in JDE (recommended to set as variable). Does not support 'All' departments.}

\item{dates_obj}{The list of dates resulting from HR_dates() function}
}
\value{
A list of data frames
}
\description{
Function to create data frames for transfers
}
\examples{
dept_transfers <- transfers(dept_HRdata, jde_dept, dates)
}
