% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{transfers_table}
\alias{transfers_table}
\title{Function to generate a table of transfers in and out of the selected department}
\usage{
transfers_table(data, dept_name)
}
\arguments{
\item{data}{The list of data frames resulting from the transfers() function}

\item{dept_name}{Name of department as it should appear in table title}
}
\value{
A kable object
}
\description{
Function to generate a table of transfers in and out of the selected department
}
\examples{
print(transfers_table(dept_transfers, dept_name = "Health"))
}
