% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{transfers_text}
\alias{transfers_text}
\title{Function to generate reporting year to date text for transfer section of report}
\usage{
transfers_text(data, rpt_date)
}
\arguments{
\item{data}{The list of data frames resulting from the transfers() function}

\item{rpt_date}{Date in reporting month}
}
\value{
A character value
}
\description{
Function to generate reporting year to date text for transfer section of report
}
\examples{
dept_transfers_text <- transfers_text(dept_transfers, dates$reporting_smonth)
}
