% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{vacancies_plot}
\alias{vacancies_plot}
\title{Function to generate plot for historical change of vacancies}
\usage{
vacancies_plot(data)
}
\arguments{
\item{data}{The list of data frames resulting from the HR_pulldata() function}
}
\value{
A ggplot object
}
\description{
Function to generate plot for historical change of vacancies
}
\examples{
print(vacancies_plot(dept_HRdata))
}
