% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HR.R
\name{vacancies_text}
\alias{vacancies_text}
\title{Function to generate text stating the number of vacant jobs and positions.}
\usage{
vacancies_text(data, rpt_enddate, dept_name)
}
\arguments{
\item{data}{The list of data frames resulting from the HR_pulldata() function}

\item{dept_name}{Name of department as it should appear in text}

\item{rpt_date}{Date in reporting month}
}
\value{
A character value
}
\description{
Function to generate text stating the number of vacant jobs and positions.
}
\examples{
dept_vacancy_text <- vacancies_text(dept_HRdata, dates$reporting_emonth, dept_name = "Health Department")
}
