% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_general.R
\name{kpi_yoy_text}
\alias{kpi_yoy_text}
\title{A function creating text for year-over-year change.}
\usage{
kpi_yoy_text(kpi_cmp, rpt_date, mode = "my")
}
\arguments{
\item{kpi_cmp}{A kpi_compare object.}

\item{rpt_date}{The first day of the reporting month.}

\item{mode}{Can be either "my" to compare to the same month in the previous year or "ytd" to compare to the same period in the previous year.}
}
\value{
A string with R markdown formatting.
}
\description{
A function creating text for year-over-year change.
}
\examples{
metric_yoy_text <- kpi_yoy_text(metric_cmp, rpt_date)
}
